# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/tenants/{tenant_name}/machines": {
            "get": {
                "description": "Listing machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machine"
                ],
                "summary": "API to list machine resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/machines.MachineListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Register new machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machine"
                ],
                "summary": "API to register new machine resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/machines.MachineRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_name}/machines/{machine_id}": {
            "get": {
                "description": "Retrieve machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machine"
                ],
                "summary": "API to retrieve machine resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/machines.MachineRetrievalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machine"
                ],
                "summary": "API to delete machine resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/machines.MachineDeletionRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updating machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machine"
                ],
                "summary": "API to update machine resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/machines.MachineUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_name}/machines/{machine_id}/actions/{action}": {
            "post": {
                "description": "Action machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machine"
                ],
                "summary": "API to perform action on machine resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/machines.MachineActionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "machines.MachineDeletionRequest": {
            "type": "object",
            "properties": {
                "machineAction": {
                    "type": "string"
                },
                "machineID": {
                    "type": "string"
                },
                "tenantName": {
                    "type": "string"
                }
            }
        },
        "machines.MachineListRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "offset": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "machines.MachineRegistrationRequest": {
            "type": "object",
            "properties": {
                "cores": {
                    "description": "The number of CPU cores for the machine.",
                    "type": "integer"
                },
                "fingerprint": {
                    "description": "The fingerprint of the machine. This can be an arbitrary string, but must be unique within the scope of the license it belongs to.",
                    "type": "string"
                },
                "hostname": {
                    "description": "The hostname of the machine.",
                    "type": "string"
                },
                "ip": {
                    "description": "The IP of the machine.",
                    "type": "string"
                },
                "license_key": {
                    "description": "The license key associated with the machine",
                    "type": "string"
                },
                "metadata": {
                    "description": "Object containing machine metadata.",
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "description": "The human-readable name of the machine.",
                    "type": "string"
                },
                "platform": {
                    "description": "The platform of the machine.",
                    "type": "string"
                }
            }
        },
        "machines.MachineRetrievalRequest": {
            "type": "object",
            "properties": {
                "machineAction": {
                    "type": "string"
                },
                "machineID": {
                    "type": "string"
                },
                "tenantName": {
                    "type": "string"
                }
            }
        },
        "machines.MachineUpdateRequest": {
            "type": "object",
            "properties": {
                "cores": {
                    "description": "The number of CPU cores for the machine.",
                    "type": "integer"
                },
                "fingerprint": {
                    "description": "The fingerprint of the machine. This can be an arbitrary string, but must be unique within the scope of the license it belongs to.",
                    "type": "string"
                },
                "hostname": {
                    "description": "The hostname of the machine.",
                    "type": "string"
                },
                "ip": {
                    "description": "The IP of the machine.",
                    "type": "string"
                },
                "license_key": {
                    "description": "The license key associated with the machine",
                    "type": "string"
                },
                "metadata": {
                    "description": "Object containing machine metadata.",
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "description": "The human-readable name of the machine.",
                    "type": "string"
                },
                "platform": {
                    "description": "The platform of the machine.",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "agg": {},
                "code": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "request_id": {
                    "type": "string"
                },
                "server_time": {
                    "type": "integer"
                }
            }
        },
...
    }
}